// WYNKWeb/src/app/Models/Customer.ts
export class Customer {
  ID: number;
  CmpID: number;
  UIN: string;
  Name: string;
  MiddleName: string;
  LastName: string;
  Address1: string;
  Address2: string;
  Address3: string;
  Location: number;
  GSTNo: string;
  MobileNo: string;
  PhoneNo: string;
  ContactPerson: string;
  IsDeleted: boolean;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy: number;
}